requirements.txt
beautifulsoup4==4.11.1
browserhistory==0.1.2
google_api_python_client==2.57.0
google_auth_oauthlib==0.5.2
lxml==4.9.1
nltk==3.7
numpy==1.23.2
pandas==1.4.3
protobuf==4.21.5
requests==2.28.1
